pyproject.toml
[project]
name = "aiohttp-openapi"
version = "0.0"
description = "Builds api schemas for aiohttp apps utilizing pydantic."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["aiohttp", "pydantic", "openapi-pydantic"]

[project.optional-dependencies]
test = ["pytest", "pytest-aiohttp", "pytest-playwright-asyncio"]

[build-system]
requires = ["setuptools >= 40.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
aiohttp_openapi = ["contrib-ui/swagger-ui/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["playwright"]
asyncio_default_test_loop_scope="session"

[tool.ruff]
line-length = 120
target-version = "py310"
lint.extend-select = [
    "I",  # isort
    "UP",  # pyupgrade
]